RavenDB-18548 moved /databases/*/hilo/next to processor
