+++ src/Raven.Server/Raven.Server.csproj
@@ -193,11 +193,6 @@
       <Link>debug.Dockerfile</Link>
     </Content>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Dashboard\Server\" />
-    <Folder Include="Documents\Handlers\Processors\BulkInsert\" />
-    <Folder Include="Documents\Sharding\Handlers\Processors\BulkInsert\" />
-  </ItemGroup>
   <PropertyGroup Condition="'$(Configuration)'=='Validate'">
     <Optimize>true</Optimize>
   </PropertyGroup>
